+++ lib/esbonio/setup.cfg
@@ -44,7 +44,7 @@ console_scripts =
     esbonio = esbonio.cli:main
 
 [options.extras_require]
-dev = black ; mypy ; flake8 ; pre-commit ; pytest ; pytest-asyncio ; pytest-cov ; pytest-timeout ; tox ; types-appdirs ; types-docutils ; types-pygments
+dev = black ; mypy ; flake8 ; pre-commit ; pytest ; pytest-lsp ; pytest-cov ; pytest-timeout ; tox ; types-appdirs ; types-docutils ; types-pygments
 
 [flake8]
 max-line-length = 88
