+++ lib/esbonio/tests/sphinx-default/workspace-src/.gitignore
