+++ lib/esbonio/tests/sphinx-default/workspace-src/.vscode/settings.json
