+++ lib/esbonio/tests/sphinx-default/workspace/.vscode/settings.json
