+++ lib/esbonio/tests/sphinx-default/workspace/_static/vscode-screenshot.png
