+++ lib/esbonio/tests/sphinx-default/workspace/glossary.rst
