+++ lib/esbonio/tests/sphinx-default/workspace/math.rst
