+++ lib/esbonio/tests/sphinx-default/workspace/theorems/pythagoras.rst
