+++ lib/esbonio/tests/sphinx-extensions/workspace/sphinx-extensions/.vscode/settings.json
